refactoring describe functioality for metadata call
